requirements.txt
asgiref==3.3.4
Django==3.2.3
docx2pdf==0.1.7
img2pdf==0.4.1
lxml==4.6.3
pikepdf==2.12.1
Pillow==8.2.0
pytz==2021.1
pywin32==227
sqlparse==0.4.1
tqdm==4.61.0
